Trampoline Revamp (#36)
Extends the functionality of the Trampoline to wrap/unwrap more Zig and
Python types.

Notably, it includes the ability to wrap/unwrap pointers to Pydust class
structs. That means e.g. returning a `*Dog` from a function.

It also replaces `__init__` for `__new__`. This better aligns with what
is actually happening. Previously with init, we needed to allocate some
memory (which was entirely undefined), then call an init function on a
class. At this point, it wasn't obvious that the class struct fields
with default values were not yet assigned.

Whereas now, the class has the responsibility of constructing a full
state struct, including defaults, that we can then assign during tp_new
to the class state.

---------

Co-authored-by: Robert Kruszewski <github@robertk.io>